// web/src/api/generated/model/employeeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { EmployeeComponentInput } from './employeeComponentInput';

export interface EmployeeInput {
  code: string;
  nameAr: string;
  nameEn: string;
  /** @nullable */
  jobTitle: string | null;
  /** @nullable */
  nationalId: string | null;
  isNational: boolean;
  joinDate: string;
  /** @nullable */
  leaveDate: string | null;
  basicSalary: number;
  /** @nullable */
  bankName: string | null;
  /** @nullable */
  bankAccount: string | null;
  /** @nullable */
  costCenterId: string | null;
  annualLeaveDays: number;
  leaveBalanceDays: number;
  /** @nullable */
  leaveBalanceDate: string | null;
  /** @nullable */
  notes: string | null;
  components: EmployeeComponentInput[];
}